#263 'mine påbegnyte søknader' gets updated after submit and 'send inn' gets rendred based on chapterCounter
