+++ src/components/SubmitButton.tsx
@@ -9,8 +9,7 @@ import {
   Grid,
 } from "@material-ui/core";
 import SendIcon from "@material-ui/icons/Send";
-import { Alert } from "@material-ui/lab";
-import React, { FC, useState } from "react";
+import { FC, useState } from "react";
 import { useHistory } from "react-router";
 import { useRecoilValue } from "recoil";
 import { firestore } from "../firebase";
@@ -21,13 +20,11 @@ import { useDocRef } from "./inputFields/saveInputFields";
 import firebase from "firebase";
 
 export const SubmitButton: FC<SubmitButtonProps> = ({
-  chapterName,
   setErrorStatus,
 }) => {
   const [open, setOpen] = useState(false);
   const currentDocID = useRecoilValue(applicationIDState);
   const currentUserID = useRecoilValue(currentUserState);
-  const [submitted, setSubmitted] = useState("in progress");
 
   const history = useHistory();
   const docRef = useDocRef();
@@ -46,10 +43,15 @@ export const SubmitButton: FC<SubmitButtonProps> = ({
       if (docData != undefined) {
         for (const application in docData.applications) {
           if (docData.applications[application].id === currentDocID) {
-            setStatusToSubmitted(docRef!, userID, application);
+            await setStatusToSubmitted(docRef!, userID, application);
             history.push("/applications");
           }
         }
+        setErrorStatus({
+          status: "error",
+          text: "Kunne ikke sende inn. Søknaden finnes ikke på brukeren.",
+        });
+        handleClose();
       } else {
         status = false;
       }
@@ -75,42 +77,40 @@ export const SubmitButton: FC<SubmitButtonProps> = ({
 
   return (
     <Grid item>
-      {chapterName === "additional" ? (
-        <Box mt={3} flexShrink={0} style={{ float: "right" }}>
-          <Button
-            variant="contained"
-            onClick={handleClickOpen}
-            endIcon={<SendIcon />}
-          >
-            Send inn
+      <Box mt={3} flexShrink={0} style={{ float: "right" }}>
+        <Button
+          variant="contained"
+          onClick={handleClickOpen}
+          endIcon={<SendIcon />}
+        >
+          Send inn
           </Button>
-          <Dialog
-            open={open}
-            onClose={handleClose}
-            aria-labelledby="alert-dialog-title"
-            aria-describedby="alert-dialog-description"
-          >
-            <DialogTitle id="alert-dialog-title">Send inn søknad</DialogTitle>
-            <DialogContent>
-              <DialogContentText id="alert-dialog-description">
-                Er du sikker på at du vil sende inn søknaden? Har du husket å
-                lagre?
+        <Dialog
+          open={open}
+          onClose={handleClose}
+          aria-labelledby="alert-dialog-title"
+          aria-describedby="alert-dialog-description"
+        >
+          <DialogTitle id="alert-dialog-title">Send inn søknad</DialogTitle>
+          <DialogContent>
+            <DialogContentText id="alert-dialog-description">
+              Er du sikker på at du vil sende inn søknaden? Har du husket å
+              lagre?
               </DialogContentText>
-            </DialogContent>
-            <DialogActions>
-              <Button onClick={handleClose} color="primary" autoFocus>
-                Gå tilbake
+          </DialogContent>
+          <DialogActions>
+            <Button onClick={handleClose} color="primary" autoFocus>
+              Gå tilbake
               </Button>
-              <Button
-                onClick={() => submitApplication(docRef!, currentUserID!.uid)}
-                color="primary"
-              >
-                Send inn
+            <Button
+              onClick={() => submitApplication(docRef!, currentUserID!.uid)}
+              color="primary"
+            >
+              Send inn
               </Button>
-            </DialogActions>
-          </Dialog>
-        </Box>
-      ) : null}
+          </DialogActions>
+        </Dialog>
+      </Box>
     </Grid>
   );
 };
